+++ src/main/java/sample/cafekiosk/spring/api/controller/order/OrderController.java
@@ -1,4 +1,4 @@
-package sample.cafekiosk.spring.api.controller;
+package sample.cafekiosk.spring.api.controller.order;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import sample.cafekiosk.spring.api.ApiResponse;
-import sample.cafekiosk.spring.domain.product.dto.request.OrderCreateRequest;
-import sample.cafekiosk.spring.domain.product.dto.OrderResponse;
+import sample.cafekiosk.spring.api.controller.order.dto.request.OrderCreateRequest;
+import sample.cafekiosk.spring.api.controller.order.dto.response.OrderResponse;
 import sample.cafekiosk.spring.api.service.order.OrderService;
 
 import java.time.LocalDateTime;
@@ -22,7 +22,7 @@ public class OrderController {
     @PostMapping("/api/v1/orders/new")
     public ApiResponse<OrderResponse> createOrder(@Valid @RequestBody OrderCreateRequest request) {
         LocalDateTime time = LocalDateTime.now();
-        return ApiResponse.of(HttpStatus.OK, "성공적으로 주문을 생성하였습니다.", orderService.createOrder(request, time));
+        return ApiResponse.of(HttpStatus.OK, "성공적으로 주문을 생성하였습니다.", orderService.createOrder(request.toServiceRequest(), time));
 
     }
 }
