+++ src/main/java/sample/cafekiosk/spring/api/service/order/dto/request/OrderCreateServiceRequest.java
@@ -9,7 +9,6 @@ import java.util.List;
 @Getter
 @NoArgsConstructor
 public class OrderCreateServiceRequest {
-    @NotEmpty(message = "상품 번호 리스트는 필수입니다.")
     private List<String> productNumbers;
 
     public OrderCreateServiceRequest(List<String> productNumbers) {
